requirements.txt
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-llms-langchain
llama-index-embeddings-langchain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq
langchain-google-genai
langchain-openai
langchain-huggingface
langchain-community
sentence-transformers
gitpython
streamlit>=1.28.0
python-dotenv
faiss-cpu
langchain-mistralai
langchain-ollama
